# Use CRTP in nodal loops