refs #31420 

The comparisons I did, running an input 10 times

Base input test/tests/nodalkernels/constant_rate/constant_rate.i -r 4
```
Without change
Average runtime: 32.410 seconds

With change
Average runtime: 32.049 seconds
```

1000x1000 only nodal variables, added std deviation output
```
without
Average runtime: 58.843401 seconds
Standard deviation: .453407 seconds

with change
Average runtime: 57.653081 seconds
Standard deviation: 2.232744 seconds
```
so overall, about 1.5% faster at most