@GiudGiud 2025-09-02
Profile for that input (no change except -r 5), before optimization
>      3.58s  1.62% 48.31%     49.34s 22.26%  ComputeNodalKernelsThread::onNode
This is relevant 

```
(pprof) top 30
Showing nodes accounting for 145.46s, 65.62% of 221.67s total
Dropped 489 nodes (cum <= 1.11s)
Showing top 30 nodes out of 158
      flat  flat%   sum%        cum   cum%
    21.33s  9.62%  9.62%    221.67s   100%  [libsasl2.2.dylib]
    10.31s  4.65% 14.27%     13.43s  6.06%  ThreadedElementLoopBase::shouldComputeInternalSide
     9.10s  4.11% 18.38%      9.43s  4.25%  libMesh::Elem::which_child_am_i
     8.95s  4.04% 22.42%     10.99s  4.96%  std::__1::__tree::__assign_multi
     8.83s  3.98% 26.40%     15.32s  6.91%  MooseVariableData::computeValuesInternal
     7.54s  3.40% 29.80%     19.55s  8.82%  libMesh::BoundaryInfo::boundary_ids
     6.56s  2.96% 32.76%      6.56s  2.96%  libMesh::fe_lagrange_1D_linear_shape
     5.03s  2.27% 35.03%      5.03s  2.27%  libMesh::Hex20::n_second_order_adjacent_vertices
        5s  2.26% 37.29%     11.56s  5.21%  (anonymous namespace)::fe_lagrange_2D_shape
     4.45s  2.01% 39.29%      4.45s  2.01%  libMesh::H1FETransformation::map_dphi
     4.36s  1.97% 41.26%      8.87s  4.00%  Kernel::computeResidual
     4.30s  1.94% 43.20%      7.83s  3.53%  TimeKernel::computeResidual
     4.12s  1.86% 45.06%      7.52s  3.39%  libMesh::FEMap::compute_affine_map
     3.62s  1.63% 46.69%     10.52s  4.75%  MooseVariableDataBase::fetchDoFValues
     3.58s  1.62% 48.31%     49.34s 22.26%  ComputeNodalKernelsThread::onNode
     3.52s  1.59% 49.89%     14.14s  6.38%  Assembly::cacheResidual
     3.24s  1.46% 51.36%      3.24s  1.46%  [libmoose-oprof.0.dylib]
     3.12s  1.41% 52.76%      7.94s  3.58%  Assembly::cacheResidualBlock
     2.92s  1.32% 54.08%      7.10s  3.20%  tc_new
     2.88s  1.30% 55.38%      2.88s  1.30%  PetscSFSetUpRanks
     2.72s  1.23% 56.61%      2.72s  1.23%  CoefDiffusion::computeQpResidual
     2.44s  1.10% 57.71%      2.44s  1.10%  PetscSortIntWithDataArray
     2.42s  1.09% 58.80%      2.43s  1.10%  dgemv_
     2.36s  1.06% 59.86%      2.36s  1.06%  libMesh::FEMap::compute_single_point_map
     2.30s  1.04% 60.90%      2.30s  1.04%  MooseMesh::getNodeBlockIds
     2.17s  0.98% 61.88%     12.69s  5.72%  MooseVariableDataBase::setActiveTags
     2.16s  0.97% 62.85%      3.90s  1.76%  libMesh::DofMap::_dof_indices
     2.08s  0.94% 63.79%    144.37s 65.13%  ThreadedElementLoopBase::operator()
     2.03s  0.92% 64.71%     14.13s  6.37%  libMesh::FE::all_shapes
     2.02s  0.91% 65.62%      3.41s  1.54%  Assembly::prepareResidual
```