@lindsayad 2025-09-10
I don't think it's worth it. We may lift CRTP from the GPU design soon based on a type erasure pattern proposed by @hugary1995